// src/app/ITS/shared/interfaces/frontend/selectedTags.ts
export interface ISelectedTag {
    "EPC_ID": string;
    "SM": string;
    "Category": string;
    "Asset_ID": string;
    "Asset_No": string;
    "Description": string;
    "Asset_LocationLocation": string;
    "LastBal": number;
    "Asset_StatusName": string;
    "Submitting_Amount": number;
    "ReceiveQty": number;
}